requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain
langchain-community
langchain-core
openai==1.3.5
pydantic==2.5.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
PyYAML==6.0.1 